% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_interface.R
\name{btree_free}
\alias{btree_free}
\title{Free free palestine}
\usage{
btree_free(tree)
}
\arguments{
\item{tree}{tree}

\item{key}{key}
}
\value{
Returns index number
}
\description{
I've no time for this.
}
