% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_interface.R
\name{btree_interval_search}
\alias{btree_interval_search}
\title{Searching for key by interval}
\usage{
btree_interval_search(tree, start, end)
}
\arguments{
\item{tree}{tree}

\item{key}{key}
}
\value{
Returns index number
}
\description{
I've no time for this.
}
