% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_interface.R
\name{btree_search}
\alias{btree_search}
\title{Searching for key}
\usage{
btree_search(tree, key)
}
\arguments{
\item{tree}{tree}

\item{key}{key}
}
\value{
Returns index number
}
\description{
I've no time for this.
}
