% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_interface.R
\name{btree_search_max}
\alias{btree_search_max}
\title{Searching minimum}
\usage{
btree_search_max(tree)
}
\arguments{
\item{tree}{tree}

\item{key}{key}
}
\value{
Returns index number
}
\description{
I've no time for this.
}
