% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_interface.R
\name{create_btree_from_df}
\alias{create_btree_from_df}
\title{Creating index}
\usage{
create_btree_from_df(df, column_name)
}
\arguments{
\item{df}{data frame}

\item{column_name}{dont'make me explain}
}
\value{
Returns c tree object
}
\description{
I've no time for this.
}
