% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liu_interface.R
\name{liu_free}
\alias{liu_free}
\title{Free LIU Index Memory}
\usage{
liu_free(index)
}
\arguments{
\item{index}{LIU index object (external pointer).}
}
\description{
Explicitly releases the memory allocated for the B+Tree index in C. 
Use this when index is no longer needed to prevent memory leaks.
}
\examples{
idx <- liu_build(mtcars, "hp")
# ...
liu_free(idx)

}
