% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liu_interface.R
\name{liu_isin}
\alias{liu_isin}
\title{Is In LIU Index}
\usage{
liu_isin(index, keys)
}
\arguments{
\item{index}{A LIU index object (external pointer).}

\item{keys}{Vector of keys (int or double must match LIU index type) to check.}
}
\value{
Vector of logical values.
}
\description{
Checks if given keys are present in LIU index. For NA it returns FALSE.
}
\examples{
\dontrun{
Checks in int index if 2, NA, 5 are present
logical <- liu_isin(idx, as.integer(c(2,NA,5)))

Checks in double index if 6.7 is present
logical <- liu_isin(idx, 6.7)
}
}
