% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liu_interface.R
\name{liu_max}
\alias{liu_max}
\title{Search for Index of Maximum Key in LIU Index}
\usage{
liu_max(index)
}
\arguments{
\item{index}{A LIU index object (external pointer).}
}
\value{
An integer vector of row indices corresponding to the maximum key.
}
\description{
Search for the largest keys in LIU index and returs their row indices
}
\examples{
# Get row indices for the smallest value in the index
# max_rows <- liu_search_max(idx)

}
