% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liu_interface.R
\name{liu_min}
\alias{liu_min}
\title{Minimum Key in LIU Index}
\usage{
liu_min(index)
}
\arguments{
\item{index}{A LIU index object (external pointer).}
}
\value{
An integer vector of row indices corresponding to the minimum key.
}
\description{
Search for the smallest key in LIU index and returs their row indices.
}
\examples{
\dontrun{
Get row indices for the smallest key in the index
min_rows <- liu_min(idx)
}
}
