% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liu_interface.R
\name{liu_search_min}
\alias{liu_search_min}
\title{Search for Minimum Key in LIU Index}
\usage{
liu_search_min(index)
}
\arguments{
\item{index}{A LIU index object (external pointer).}
}
\value{
An integer vector of row indices corresponding to the minimum key.
}
\description{
Search for the smallest keys in LIU index and returs their row indices
}
\examples{
# Get row indices for the smallest value in the index
# min_rows <- liu_search_min(idx)

}
