% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liu_interface.R
\name{liu_search_prefix}
\alias{liu_search_prefix}
\title{Prefix Search in String LIU Index}
\usage{
liu_search_prefix(index, prefix)
}
\arguments{
\item{index}{A LIU index object (external pointer).}

\item{prefix}{Numeric scalar, beginning of the range (inclusive). Must}
}
\value{
Integer number of row indices with keys within the range.
}
\description{
Count all row indices with keys within a specified numerical range [start, end) in LIU. 
index. This operation is very efficient due to the B+Tree structure. You can leave
the start, the end, or both blank for an unbounded count.
}
\examples{
\dontrun{
Count rows where 10 <= key < 50
rows <- liu_count(idx, 10, 50)

Count row indices with keys greater or equal to 2.5
rows <- liu_count(idx, 2.5)
}
}
