% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liu_interface.R
\name{liu_sum_range}
\alias{liu_sum_range}
\title{Range Sum in LIU Index}
\usage{
liu_sum_range(df, index, start = NA, end = NA)
}
\arguments{
\item{df}{Data frame on which given index is built.}

\item{index}{A LIU index object (external pointer).}

\item{start}{Numeric scalar, beginning of the range (inclusive). Must
match index type (int or double).}

\item{end}{Numeric scalar, end of the range (exclusive). Must
match index type (int or double).}
}
\value{
Integer vector of row indices for values within the range.
}
\description{
Sum all values within a specified numerical range [start, end) in LIU. 
index.
}
\examples{
Sum all prices where 10 <= price < 50
idx <- liu_build(df, "price")
liu_sum_range(df, idx, 10, 50)

}
