% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_interface.R
\docType{package}
\name{liulibrary-package}
\alias{liulibrary}
\alias{liulibrary-package}
\title{liulibrary: LIU Library}
\description{
Implements SQL-like indexes for data frames
}
\author{
\strong{Maintainer}: Marek Skok \email{marekskok@pw.edu.pl}

}
