% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liu_interface.R
\docType{package}
\name{liu-package}
\alias{liu}
\alias{liu-package}
\title{liu: Lightning Index Units}
\description{
Implements SQL-like indexes for data frames. Makes searching for keys or merging much faster.
}
\author{
\strong{Maintainer}: Marek Skok \email{marekskok@pw.edu.pl}

}
