% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liu_interface.R
\name{liu_build}
\alias{liu_build}
\title{Build LIU Index}
\usage{
liu_build(df, column_name)
}
\arguments{
\item{df}{A data frame to be indexed.}

\item{column_name}{A character string specifying the column to index. 
(int or double column)}
}
\value{
C LIU index object
}
\description{
This function if foundation of LIU package. It builds a high-performance 
SQL-like B+Tree index for a data frame column, enabling fast joins and lookups.
Works only for columns of ints or doubles.
}
\examples{
idx <- liu_build(mtcars, "mpg")

}
