% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liu_interface.R
\name{liu_join}
\alias{liu_join}
\title{Fast Inner Join}
\usage{
liu_join(df_left, column_name, df_right, index, how = "inner")
}
\arguments{
\item{df_left}{Data frame (left side of the join).}

\item{column_name}{Name of the join numeric column (must exist in left data frames).}

\item{df_right}{The "indexed" data frame (right side of the join).}

\item{index}{A LIU index object built on the join column of `df_right`.}

\item{how}{A character string "inner" or "left".}
}
\value{
Returns merged data frame.
}
\description{
Performs a high-performance Join between two data frames using a LIU index. 
For now only inner (default) and left join are available. It takes 4
arguments left data table and name of one of its columns, right data
table and index built on it. Type of chosen column must match index type.
}
\examples{
idx <- liu_build(df_b, "id")
merged <- liu_join(df_a, "id", df_b, idx)

idx <- liu_build(df_b, "id")
merged <- liu_join(df_a, "id", df_b, idx, "left")

}
