% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liu_interface.R
\name{liu_search}
\alias{liu_search}
\title{Search for Keys in LIU Index}
\usage{
liu_search(index, key)
}
\arguments{
\item{index}{LIU index object (external pointer).}

\item{key}{Vector of keys (int or double must match LIU index type) to search for.}
}
\value{
Integer vector of row indices where the keys were found. 
Returns an empty vector if none of the keys exis.
}
\description{
Performs a fast lookup in the LIU index to find all row indices associated 
with given vector of keys.
}
\examples{
row_idx <- liu_search(idx, 110)
mtcars[row_ids, ]

row_ids <- liu_search(idx, c(6.7,21.15))
df[row_idx]

}
