% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liu_interface.R
\name{liu_search_range}
\alias{liu_search_range}
\title{Range Search in LIU Index}
\usage{
liu_search_range(index, start = NA, end = NA)
}
\arguments{
\item{index}{A LIU index object (external pointer).}

\item{start}{Numeric scalar, beginning of the range (inclusive). Must
match index type (int or double).}

\item{end}{Numeric scalar, end of the range (exclusive). Must
match index type (int or double).}
}
\value{
Integer vector of row indices for values within the range.
}
\description{
Finds all row indices with values within a specified numerical range [start, end) in LIU. 
index. This operation is very efficient due to the B+Tree structure. You can leave
the start, the end, or both blank for an unbounded search.
}
\examples{
Find rows where 10 <= key < 50
rows <- liu_search_range(idx, 10, 50)
df[rows]

Row indices with keys greater or equal to 2.5
rows <- liu_search_range(idx, 2.5)

}
